+++ src/components.d.ts
@@ -52,10 +52,6 @@ export namespace Components {
          */
         "disabled": boolean;
         /**
-          * The validation error text
-         */
-        "errorText"?: string;
-        /**
           * The help or explanation text
          */
         "helpText"?: string;
@@ -20079,10 +20075,6 @@ declare namespace LocalJSX {
           * Disables the checkbox via HTML disabled attribute. Checkbox takes on a distinct visual state. Cursor uses the not-allowed system replacement and all keyboard and mouse events are ignored.
          */
         "disabled"?: boolean;
-        /**
-          * The validation error text
-         */
-        "errorText"?: string;
         /**
           * The help or explanation text
          */